+++ python/mujoco/structs.cc
@@ -1186,6 +1186,8 @@ This is useful for example when the MJB is not available as a file on disk.)"));
   X(objid);
   X(category);
   X(matid);
+  X(texid);
+  X(texuniform);
   X(texcoord);
   X(segid);
   X(emission);
@@ -1202,6 +1204,7 @@ This is useful for example when the MJB is not available as a file on disk.)"));
   X(pos);
   X(mat);
   X(rgba);
+  X(texrepeat);
 #undef X
 
   DefinePyStr(mjvGeom, "label", &raw::MjvGeom::label);